requirements.txt
aiofiles==23.2.1
annotated-types==0.6.0
anyio==4.3.0
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
async-timeout==4.0.3
asyncpg==0.29.0
bcrypt==4.1.2
cffi==1.16.0
click==8.1.7
cryptography==42.0.5
dnspython==2.6.1
email_validator==2.1.1
exceptiongroup==1.2.1
fastapi==0.110.3
fastapi-users==13.0.0
greenlet==3.0.3
h11==0.14.0
idna==3.7
makefun==1.15.2
pwdlib==0.2.0
pycparser==2.22
pydantic==2.7.1
pydantic_core==2.18.2
PyJWT==2.8.0
python-dotenv==1.0.1
python-multipart==0.0.9
sniffio==1.3.1
SQLAlchemy==2.0.29
starlette==0.37.2
typing_extensions==4.11.0
uvicorn==0.29.0
